api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["NFT"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.3"
djangorestframework = "^3.12.4"
Pillow = "^8.2.0"
python-dotenv = {extras = ["cli"], version = "^0.17.1"}
six = "^1.16.0"
psycopg2 = "^2.8.6"
django-environ = "^0.4.5"
django-imagekit = "^4.0.2"
django-allauth = "^0.44.0"
nanoid = "^2.0.0"
django-cors-headers = "^3.7.0"
gunicorn = "^20.1.0"
uvicorn = "^0.14.0"
django-storages = {extras = ["boto3"], version = "^1.11.1"}
django-extensions = "^3.1.3"
drf-nested-routers = "^0.93.3"
django-filter = "^2.4.0"
django-rest-knox = "^4.1.0"


[tool.poetry.dev-dependencies]
black = "^21.6b0"
ipython = "^7.25.0"
isort = "^5.9.1"
pylint = "^2.9.3"
pylint-django = "^2.4.4"
Werkzeug = "^2.0.1"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
